fix: ensure fix-volumes uses node user on Mac
Using `stat -c "%u:%g" .` to determine the UID:GID does
not work when the host machine is Mac OSX and the
working directory (`.`) is itself a volume linked to the
host machine. On Mac, the command will always show
root (0:0) as the owner because all volumes are always
owned by root. This is true even if the underlying image
has this directory pre-created with different ownership.
The volume ownership masks the ownership built into
the image.

Since the true purpose of `fix-volumes.sh` is to ensure
that all volumes are owned by the same user who will
run the command, my fix is to pass this user name into
the script as an argument. Currently we could assume
"node" user, but the script is more versatile if it accepts
the user name as an argument because only the calling
Dockerfile/entrypoint truly knows which user it will be
running the command as.

Signed-off-by: Eric Dobbertin <eric@dairystatedesigns.com>